{python3Packages.}pwndbg: remove
pwndbg has very specific versioning requirements for its dependencies,
such as `unicorn`. Due to the amount and complexity of the packages,
it's not feasible to maintain pinned versions just for `pwndbg`, as
they'd come with a massive overhead of patches and build changes that'd
need to be overridden here. As `pwndbg` already provides a downstream
flake with all that versioning, users should rather consume it from
there.